Generalize transport layer port to TransportPacket and TransportPacket.header
