+++ pcap4j-core/src/main/java/org/pcap4j/packet/TransportPacket.java
@@ -1,0 +1,29 @@
+package org.pcap4j.packet;
+
+import org.pcap4j.packet.namednumber.Port;
+
+/**
+ * @author Ferran Altimiras
+ */
+public interface TransportPacket extends Packet {
+
+  @Override
+  public TransportHeader getHeader();
+
+  /**
+   * The interface representing the Transport layer packet's header.
+   *
+   * @author Ferran Altimiras
+   */
+  public interface TransportHeader extends Header {
+	/**
+	 * @return Source port
+	 */
+	public Port getSrcPort();
+
+	/**
+	 * @return Destination port
+	 */
+	public Port getDstPort();
+  }
+}
